working gui

+++ main.py
@@ -5,19 +5,23 @@ import audio
 from recognizer.data_handler import DataHandler
 from plotter import Plotter
 from mapper import Mapper
+from demo import Gui
 
-LOW_FREQ = 18000
-HIGH_FREQ = 18500
+LOW_FREQ = 20000
+HIGH_FREQ = 20500
 IN_BUFFER_SIZE = 2048
 SAMPLE_RATE = 44100
 FREQ_BIN_RES = SAMPLE_RATE / (IN_BUFFER_SIZE * 2)
 
 
-def end_program(processes):
-    print("\nEnding processes...")
-    for p in processes:
-        p.terminate()
-    print("Program has ended.")
+def check_for_end_program(processes, end_program_q):
+    #print("\nEnding processes...")
+    while True:
+        if not end_program_q.empty() or keyboard.is_pressed('esc'):
+            for p in processes:
+                p.terminate()
+            break
+    #print("Program has ended.")
 
 
 if __name__ == "__main__":
@@ -27,6 +31,7 @@ if __name__ == "__main__":
     result_queue = mp.Queue()
     recognition_queue = mp.Queue()
     action_queue = mp.Queue() # For gesture-action mapping
+    end_program_q = mp.Queue()
 
     output_tones = audio.Output(low_freq=LOW_FREQ, high_freq=HIGH_FREQ, 
                                 sample_rate=SAMPLE_RATE)
@@ -35,6 +40,7 @@ if __name__ == "__main__":
                            buffer_size=IN_BUFFER_SIZE, sample_rate=SAMPLE_RATE, 
                            peak_freqs=(LOW_FREQ, HIGH_FREQ), 
                            f_bin_res=FREQ_BIN_RES)
+    """
     velocity_analyzer = velocity.VelocityAnalyzer(f_domain_q=velocity_queue, 
                                                   v_plot_q=velocity_plot_queue, 
                                                   peak_freqs=(LOW_FREQ, HIGH_FREQ), 
@@ -43,21 +49,29 @@ if __name__ == "__main__":
     plotter = Plotter(in_plot_q=input_plot_queue, v_plot_q=velocity_plot_queue,
                       res_q=result_queue, low_freq=LOW_FREQ, high_freq=HIGH_FREQ, 
                       f_bin_res=FREQ_BIN_RES)
+    """
+    
+                      
     recognizer = DataHandler(data_q=recognition_queue, res_q=result_queue,
                              action_q=action_queue)
     mapper = Mapper(action_q=action_queue)
+    demo_gui = Gui(res_q=result_queue, end_program_q=end_program_q)
 
     output_p = mp.Process(target=output_tones.start)
     input_p = mp.Process(target=audio_in.get)
+    """
     get_velocity_p = mp.Process(target=velocity_analyzer.get_v)
     plot_p = mp.Process(target=plotter.plot_all)
+    """
+
+    
     recognition_p = mp.Process(target=recognizer.process_data)
     mapper_p = mp.Process(target=mapper.state_machine)
+    demo_gui_p = mp.Process(target=demo_gui.check_for_gesture)
 
-    processes = [output_p, input_p, get_velocity_p, recognition_p, plot_p]#, mapper_p]
+    processes = [output_p, input_p, recognition_p, demo_gui_p]#, plot_p]#, get_velocity_p], mapper_p]
 
     for p in processes:
         p.start()
 
-    keyboard.wait('esc')
-    end_program(processes)
+    check_for_end_program(processes, end_program_q)
